fix bugs, change mirrors

+++ zeppelin-zengine/src/main/java/org/apache/zeppelin/helium/HeliumBundleFactory.java
@@ -385,6 +385,9 @@ public class HeliumBundleFactory {
     }
 
     WebpackResult result = getWebpackResultFromOutput(bundleStdoutResult);
+    if (result == null) {
+      return heliumBundle;
+    }
     if (result.errors.length > 0) {
       FileUtils.deleteQuietly(heliumBundle);
       throw new IOException(result.errors[0]);
